riscv: implement VirtIO target
This allows running RISC-V tests in CI using QEMU, which should help
catch bugs.
