+++ targets/riscv-qemu.json
@@ -1,0 +1,7 @@
+{
+	"inherits": ["riscv"],
+	"features": ["+a", "+c", "+m"],
+	"build-tags": ["virt", "qemu"],
+	"linkerscript": "targets/riscv-qemu.ld",
+	"emulator": ["qemu-system-riscv32", "-machine", "virt", "-nographic", "-bios", "none", "-kernel"]
+}
